+++ core/src/test/java/com/fujitsu/dc/test/jersey/bar/BarInstallUserDataLinksLimitTest.java
@@ -69,7 +69,7 @@ public class BarInstallUserDataLinksLimitTest extends JerseyTest {
     private static final String REQ_CONTENT_TYPE = "application/zip";
     private static final String REQUEST_NORM_FILE = "bar-install.txt";
     private static final String RESOURCE_PATH = "requestData/barInstall";
-    private static final String DEFAULT_SCHEMA_URL = "https://fqdn/testcell1";
+    private static final String DEFAULT_SCHEMA_URL = "https://fqdn/testcell1/";
     private static final String MASTER_TOKEN_NAME = Setup.MASTER_TOKEN_NAME;
 
     private static int linkNtoNMaxSize = 40;